Names n colors (#15)
* new hold key

change the hold key from c to h because that makes more sense. i hope i dont have to go back and change this to getup to spec.

* new colors

updated the colors for the tritris and tetris pieces

* update tritris cases

using the new cases we agree on to map to all the new omni tris pieces. we'll follow this new pattern for all future pieces