Bitris piece (#17)
* new colors

updated the colors for the tritris and tetris pieces

* update tritris cases

using the new cases we agree on to map to all the new omni tris pieces. we'll follow this new pattern for all future pieces

* add bitris piece

added the new piece and its new color. I had to change the piece cases to becasue the lowercase i case was previously used for a tetris piece.

* remove conflict with piece cases